[ethif] add the ability to capture packets sent to/received from the network
Packets read or written are pushed to an 'Lwt_bounded_stream' which will drop
rather than block when the stream is full. The default size of the
stream is 0, i.e. all packets are dropped i.e. capturing is disabled.

Clients which wish to process the packets should change the default stream
bound to more than 0 and then use "nget" in a blocking thread.