Performance tweak - removed some unnecessary signalling between tcp threads
Previously every tcp ack received was being signalled to the TX thread
and since an mvar was being used for the signalling, the RX thread would
block till the TX side had processed the previously signalled ACK.  Fixed,
and also added a test to only signal if the ack has info that matters
to the TX thread.  Gives about a 15% speedup.