Expose IPV4 through the STACKV4 interface.
The socket version is just a skeleton that fails if most functions
are used.  These can eventually be filled in with the UNIX
equivalents, but I'm deferring this work as it's hard to get it
done quickly and portably, and noone needs this functionality
in the socket backend at present.

This does allow the direct networking stack to transmit and send
raw IPv4 packets, as well manipulate the IP address bindings on
the interface.

Requires the associated mirage-types change to compile