Fully process the last ACK in 3way handshake for server connections.
Invoke input-handling code on the last ACK received as part
of a 3-way handshake, before invoking the user-defined callback for
flow processing.  If these operations are done the other
way 'round, the connection remains in state Syn_rcvd until the
retransmission timer expires and a duplicate SYN/ACK is sent,
causing the remote end to send a duplicate ACK.  This introduces an
unnecessary delay and can result in inconsistent internal state under
some use cases.