Added checksum_stubs.c from mirage-platform
This is only used for networking, and isn't specific to a Mirage
platform (the Xen and Unix versions should be the same).

We now always use the x86_64-specific optimised version from mirage-xen
on x86_64 (whether Unix or Xen), and always use the generic version from
mirage-unix on other platforms.