Allocate frame in `nd_ns_output'
This required shuffling code around quite a bit and in particular
removing the module `Ndpv6' and `Routing' from the picture, leaving a
single, monolithic module `Ipv6'.

The reason is that IPv6, ICMPv6, and NDPv6 are actually all protocols
lying on the same layer and there are a lot of cyclic dependencies
between them.  Later on it will be necessary to figure out a way to
decouple the different modules but for now it is simpler to put
everything in the same namespace.