Make core logic independent of Lwt
When necessary a function now returns a pair [(pkts, timers)] consisting
of a list of ethernet packets that should be written to the underlying
network interface and a list of floats representing number of seconds
until the `tick' function should be invoked.  This way we do not have to
do busy waiting when nothing is going on. The timers are managed
automatically by the Lwt/Mirage frontend (functor Make).