ARP now handles ARP frames, not Ethernet frames with ARP payload
ARP until now received full ethernet frames, and defined ARP packets
in a way such that they include ethernet frames. This prevents code
reuse and is a discrepancy of the standard.

Furthermore, the data of the Ethernet header was never touched, and
the fix for assembling an ARP frame is trivial.