wrap ports for listeners at 65536
Previously, it was possible to register listeners for ports > 65535.
This had unexpected results, as serializing the requested port number
elsewhere in the stack would result in truncating the port number, but
reply packets would not match any known listener (since the registered
listener was at the requested port number).  Register the given port
number mod 65536 to avoid this even if the user submits an unreasonable
port number, for consistency with other parts of the stack.