remove count from Fin_wait_2
Previously, the state machine incremented a counter on
subsequent ACKs received when in the Fin_wait_2 state.  The logic around
this counter would only move the connection out of Fin_wait_2 if the
counter had not increased beyond the initial count, meaning the
connection would never terminate if duplicate ACKs were received.  The
counter logic is removed in this patch.