replace Reset state with Closed; check ack # for RSTs
For connections that have an established Pcb, Segment.Rx.input will
check the incoming sequence number against the window state.  Therefore,
we don't need to check this in State also.  Make Segment a bit more
rigorous by checking that the sequence number for incoming RSTs is at
the left edge of the window (per RFC 5961).