fix #162 (duplicate ACKs during closure)
Signaling to the thread that "spools the data into an application
receive buffer" was incorrectly being treated equally between the situation
where we need to acknowledge data but have none to send ourselves, and
the situation where the connection is being closed.  This resulted in a
duplicate ACK being sent after any ACK to a FIN/ACK segment.  Instead,
check to see whether the connection is being closed; if it is, don't
cause another ACK to be sent early.