remove unused code, including convenience fns for echo & would_fragment
Rather than a pattern where the user submits a set of arguments for
constructing the packet, Marshal functions generally expect to take a
`t` argument now which the caller has constructed with the relevant
information.  Remove the `echo_request`, `echo_reply`, and
`would_fragment` functions in favor of this pattern.