replace calls to OS.Time with calls to Vnetif_common.Time
Some test code was relying on the OS module, linked in when
mirage-console.unix was provided, to indirectly provide access to the
`sleep` function.  Removing `mirage-console.unix` requires the tests to
access `sleep` in some other way.  `Vnetif_common` already provides an
implementation of `V1_LWT.TIME` (using Lwt_unix), so use that.