Support IP fragmentation
Uses a fixed mtu of 1500, as the network
layer currently doesn't give a way to ask it.

Ipv4 write/writev should have an optional argument
for specifing if fragmentation is allowed, that can
be used for TCP Path MTU Discovery.

The Ipv4 functor now requires a Time parameter as well,
so need to update the scripts that do the network stacks
config for mirage.