Fix race when closing test pcap files
As the backend listeners we use to capture the packets are called in
async, the pcap file could be closed before the final packets had been
recorded. With this change we use the new `unregister_and_flush`
function available in mirage-vnetif 0.2.0 and later to make sure all the
listener callbacks for the client that is unregistering have returned
before closing the file.

Signed-off-by: Magnus Skjegstad <magnus@skjegstad.com>