Changed test for initial rto to check the value that is better for the mirage stack
The initial RTO is recommended to be set to 3s but then as soon as the
first RTT is measured the correct RTO should get set with 1s as the
minimum.  So the effective real initial RTO is usually 1s.
This intention needs to be adapted for the mirage tcp timer.
The tcp timer works as follows: when it goes off it checks if the
pkt it was scheduled to rexmit has been acked - if not it rexmits it,
then if there are un-acked packets it takes the last un-acked packet
and schedules itself to rexmit it.  This means that on average for
a rexmit to happen two cycles of the timer need to go off.  So to
get a 1s RTO on average, the timer needs to be set to 2/3 seconds.
Also because of the way the timer works having the first RTO set
to be 3s means that we are forced to suffer it - it is much better
to set it to be 1s directly. Hence this PR and this fix to the test.