port to jbuilder
This significantly rearranges the source tree, but preserves
the old META format as much as possible.  It adds a new `tcpip.unix`
which is needed for the unix checksum stubs to be linked in.
Previously those were part of the `tcpip` library, which risks
a double link in the case of the xen and solo5 variants.

Build time seems to be improved by 4x with this change...