icmpv4_socket: correct the IPv4 total_length field
On some OSes (particularly macOS) the IP header returned to userspace
has a bogus total length (e.g. 16384). This probably reflects an internal
kernel buffer size rather than the on-the-wire size of the original
packet.

This patch corrects the IPv4 header total length so that the IPv4 parser
can parse the result. Previously it would return an Error.

Signed-off-by: David Scott <dave@recoil.org>