Fix TCP deadlock with slow receivers: RX window closed and never opens again.
When the User_buffer.Rx queue is full, the RX window normally closes.
When the data is eventually consumed, no ACKs are sent (even though
there is now free space in the buffer), and the connection hangs.

We fix this by making sure the window is at least 3 MSS (with some care
to cope with window scaling): this way, there will be some unACKed data
not yet queued in the user buffer when the latter is full. When
data is consumed and the segments can be queued in the user buffer,
ACKs will be sent and the window will reopen progressively.