introduce ipv4/Fragments.fragment : ~mtu:int -> Ipv4_packet.t -> Cstruct.t -> Cstruct.t list
a function that takes the MTU, an IPv4 header, and a payload, and returns a list
of IPv4 segments to be sent out. This is to-be-called with the remaining payload
after the first segment. Use this function in Static_ipv4.write.

For what its worth: >95% of IPv4 packets are not fragmented, i.e. it is crucial
that the fast path (no fragmentation) does least allocation!
In contrast to earlier code, this now copies the payload (Cstruct.concat rest)
before passing to Fragments.fragment (which copies once more
(Cstruct.append hdr payload), and then the call to writeout blits it into the
network-device allocated storage).

It is also slightly incorrect (in the more general setting where
Fragments.fragment is used from elsewhere with an Ipv4_packet.t that contains
options (which shouldn't be copied)) (requires a different representation of
options -- a list of (bool * Cstruct.t) where the bool is the first bit
(indicating whether to copy the option (or not) into following IPv4 fragments)).

//cc @yomimono @linse @cfcs
https://github.com/mirage/mirage-nat/issues/31
https://github.com/mirage/qubes-mirage-firewall/issues/73