Fix ipv6 test exception
The ipv6 test disconnected from the vnetif backend while the stack was
still running, resulting in a `Not_found` exception being returned from
mirage-vnetif when Ethernet.write was called. This only became a problem
in a recent commit that fixed an issue that prevented some of the
background timers from running in the IPv6 stack.

The V.disconnect call seems to have been added to work around another
IPv6 issue where the UDP sender will queue packets with an empty source
address until an IPv6 address is configured. This results in multiple
packets with empty source address being received after the first correct
packet, unless the receiver disconnected from the backend before this
happens.

This updates the test to not send UDP packets until an IP address is
acquired and removes the call V.disconnect in the receiver. It does not
change the queuing behaviour of the IPv6 stack, which may also be
incorrect.