Fix MTU enforced test backend
The previous backend contained a ref to the MTU in the module, so the
MTU was remembered between tests. The iperf test would not succeed
unless the mtu+tcp test had run previously, as it set the MTU to 9000
instead of the expected 1500.

This renames the MTU enforced backend to Frame_size_enforced to make it
clear that the size limit is on the Ethernet layer. The frame size is
stored in the returned `t` and is now reset between tests. A helper
function is also added to set the enforced frame size based on the IP
layer MTU (MTU + size of ethernet header).

This also re-enables the iperf test as it now succeeds. The previous
failure was due to the enforced frame size being equal to the MTU,
leaving no room for the Ethernet header.