Fix iperf test to report results in mbps, run longer
This updates the iperf test to report speed in mbit/s once every second
and when the test ends. The output can be seen when alcotest is executed
in verbose mode.

Also increases the length of the slow tests to around ~100MB per
transfer. These are not run in CI by default.