Add iperf ipv6 test
This duplicates the iperf test for ipv6. As the V4 stack signatures
contain a reference to the IP version (e.g. listen_tcpv4 etc) it was
difficult to reuse the test without major changes. We should clean this
up when we unify the signatures to avoid duplicated tests for IPv4/IPv6.